Factor out the command-line argument processing code from widgets_overview.py into a separate module which all of the example programs can now reference to have nice command-line interfaces.
